feat(api): allow creation of bundles with opentrons_simulate
opentrons_simulate now lets you specify directories in which to find data and
labware to use in your protocol. It also lets you specify --bundle to output a
protocol bundle for later execution.

For instance, given a directory structure like this:

my_protocol/
  - my_protocol.ot2.py
  - my_labware.json
  - my_other_labware.json
  - my_data.csv

then going into my_protocol/ and doing
opentrons_simulate -D my_protocol.ot2.py --bundle

will create my_protocol.ot2.zip, a protocol bundle containing the python
protocol, the labware definitions, and the data file. This can then be executed
using opentrons_execute or the Opentrons App, and the labware definitions and
data files are accessible to the protocol.

This feature is only accessible if the API v2 feature flag is turned on.